+++ pyfgaws/core/tests/test_core.py
@@ -1,11 +1,11 @@
-"""Tests for :module:`~pyfgaws.util`"""
+"""Tests for :module:`~pyfgaws.core`"""
 
-from pyfgaws.util import add_jitter
+from pyfgaws.core import add_jitter
 import pytest
 
 
 def test_add_jitter() -> None:
-    for delay in [1, 5, 10, 5.5]:
+    for delay in [-10, 0, 1, 5, 10, 5.5]:
         for width in [1, 5, 10, 2.5]:
             for minima in [0, 1, 5, 10, 0.5]:
                 for _ in range(15):
@@ -17,10 +17,6 @@ def test_add_jitter() -> None:
 
 
 def test_add_jitter_bad_args() -> None:
-    with pytest.raises(AssertionError):
-        add_jitter(delay=0)
-    with pytest.raises(AssertionError):
-        add_jitter(delay=-10)
     with pytest.raises(AssertionError):
         add_jitter(width=0)
     with pytest.raises(AssertionError):
